msync/requirements.txt
anytree==2.8.0
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.3.1
asn1crypto==1.4.0
backcall==0.2.0
bauh==0.9.10
beautifulsoup4==4.9.3
browser-cookie3==0.12.0
browsercookie==0.7.7
btrfsutil==5.9
CacheControl==0.12.6
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.3.1
cupshelpers==1.0
decorator==4.4.2
distlib==0.3.1
distro==1.5.0
Django==3.1.5
djangorestframework==3.12.2
filelock==3.0.12
future==0.18.2
gestures==0.2.3
Glances==3.1.5
gufw==20.4.0
html5lib==1.1
idna==2.10
importlib-metadata==3.3.0
ipython==7.19.0
ipython-genutils==0.2.0
isc==2.0
jedi==0.17.2
jeepney==0.6.0
keyring==21.8.0
layoutswitcherlib==0.8.16
LibAppArmor==3.0.1
libvirt-python==6.4.0
lockfile==0.12.2
louis==3.16.0
lxml==4.6.2
lz4==3.1.1
mallard-ducktype==1.0.2
more-itertools==8.6.0
msgpack==1.0.0
netsnmp-python==1.0a1
npyscreen==4.10.5
OpenCC==1.1.1
ordered-set==4.0.2
packaging==20.7
pacman-mirrors==4.19.1
parso==0.7.1
pbkdf2==1.3
pep517==0.9.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
ply==3.11
progress==1.5
prompt-toolkit==3.0.8
psutil==5.7.3
ptyprocess==0.6.0
pwquality==1.4.4
pyaes==1.6.1
pyaml==20.4.0
pycairo==1.20.0
pycparser==2.20
pycryptodome==3.9.9
pycups==2.0.1
pycurl==7.43.0.6
pyenchant==3.1.1
Pygments==2.7.3
PyGObject==3.38.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pytoml==0.1.21
pytz==2020.5
pyxdg==0.26
PyYAML==5.3.1
ranger-fm==1.9.3
requests==2.25.1
resolvelib==0.5.3
retrying==1.3.3
rpm==4.16.0
SecretStorage==3.3.0
selenium==3.141.0
sip==4.19.24
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
stevedore==3.3.0
team==1.0
toml==0.10.2
traitlets==5.0.5
ufw==0.36
urllib3==1.25.10
vboxapi==1.0
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.0
